+++ Processing/LetterInfo.cs
@@ -11,6 +11,8 @@ namespace PLGL.Processing
     {
         public Letter Letter { get; set; }
         public SyllableInfo Syllable { get; set; }
+        public LetterGroup Group { get; set; }
+
         public LetterInfo AdjacentLeft { get; set; }
         public LetterInfo AdjacentRight { get; set; }